+++ SqlCore/Install/Notifications/CreateTables/__CREATE_NOTIFICATION_EVENT_CONSUMERS_TABLE__.SqlServer.sql
@@ -1,7 +1,7 @@
 CREATE TABLE t_notification_event_consumers
 (
-	id_not_evnt_consumer int NOT NULL,
-	id_notification_event int NOT NULL,
+	id_not_evnt_consumer bigint NOT NULL,
+	id_notification_event bigint NOT NULL,
 	id_acc int NOT NULL,
 	dt_crt datetime NOT NULL,
   CONSTRAINT PK_t_Notification_Event_Consumers PRIMARY KEY CLUSTERED(id_not_evnt_consumer ASC),
